+++ packages/collector/src/logger.js
@@ -1,3 +1,5 @@
+/* eslint-disable dot-notation */
+
 'use strict';
 
 var bunyan = require('bunyan');
@@ -35,8 +37,12 @@ exports.init = function(config, isReInit) {
       stream: bunyanToAgentStream,
       level: 'info'
     });
-    if (config.level) {
+    if (process.env['INSTANA_DEBUG'] || process.env['INSTANA_DEV']) {
+      parentLogger.level('debug');
+    } else if (config.level) {
       parentLogger.level(config.level);
+    } else if (process.env['INSTANA_LOG_LEVEL']) {
+      parentLogger.level(process.env['INSTANA_LOG_LEVEL'].toLowerCase());
     }
   }