+++ packages/collector/src/uncaught/index.js
@@ -36,7 +36,6 @@ exports.init = function(_config, _downstreamConnection, _processIdentityProvider
   config = _config;
   downstreamConnection = _downstreamConnection;
   processIdentityProvider = _processIdentityProvider;
-  setDefaults();
   if (config.reportUncaughtException) {
     if (!processIdentityProvider) {
       logger.warn('Reporting uncaught exceptions is enabled, but no process identity provider is available.');
@@ -48,17 +47,6 @@ exports.init = function(_config, _downstreamConnection, _processIdentityProvider
     }
   }
 };
-
-function setDefaults() {
-  if (config.tracing && config.tracing.stackTraceLength != null) {
-    stackTraceLength = config.tracing.stackTraceLength;
-  }
-  config.reportUncaughtException = config.reportUncaughtException === true;
-  // Make reportUncaughtException implies reportUnhandledPromiseRejections unless explicitly disabled.
-  if (config.reportUnhandledPromiseRejections == null) {
-    config.reportUnhandledPromiseRejections = config.reportUncaughtException;
-  }
-}
 
 exports.activate = function() {
   activateUncaughtExceptionHandling();