+++ packages/core/src/index.js
@@ -1,6 +1,7 @@
 'use strict';
 
 var log = require('./logger');
+var normalizeConfig = require('./util/normalizeConfig');
 
 // Require this first to ensure that we have non-instrumented http available.
 var uninstrumentedHttp = require('./uninstrumentedHttp');
@@ -16,6 +17,8 @@ module.exports = exports = {
 
 exports.init = function(config, downstreamConnection, processIdentityProvider) {
   log.init(config);
+  config = normalizeConfig(config);
+  exports.secrets.init(config);
   exports.util.requireHook.init(config);
   exports.tracing.init(config, downstreamConnection, processIdentityProvider);
 };